Refactor: API 호출 로직 및 type 분리
