+++ src/hooks/queries/useGetCommentQuery.ts
@@ -1,20 +1,7 @@
-import axiosInstance from "@/services/axios/axiosInstance";
 import { useQuery } from "@tanstack/react-query";
-import { Comment } from "@/types/comment";
+import { getComments } from "@/api/comment";
 
-// 댓글 목록 조회 api
-export const getComments = async (playlistId: string): Promise<Comment[]> => {
-  const response = await axiosInstance.get<Comment[]>("/comment", {
-    params: {
-      playlist_id: `eq.${playlistId}`,
-      select: "*,user:author_id(nickname, profile_image)",
-      order: "created_at.desc",
-    },
-  });
-
-  return response.data;
-};
-
+// 댓글 조회 쿼리
 export const useGetCommentQuery = (playlistId: string) => {
   return useQuery({
     queryKey: ["comments", playlistId],
